ArduSub: adjust for rangefinders getting longer ranges

+++ ArduSub/Sub.h
@@ -148,8 +148,8 @@ private:
         bool enabled:1;
         bool alt_healthy:1; // true if we can trust the altitude from the rangefinder
         int16_t alt_cm;     // tilt compensated altitude (in cm) from rangefinder
-        int16_t min_cm;     // min rangefinder distance (in cm)
-        int16_t max_cm;     // max rangefinder distance (in cm)
+        int32_t min_cm;     // min rangefinder distance (in cm)
+        int32_t max_cm;     // max rangefinder distance (in cm)
         uint32_t last_healthy_ms;
         float inertial_alt_cm;                  // inertial alt at time of last rangefinder sample
         float rangefinder_terrain_offset_cm;    // terrain height above EKF origin
